refactor(orchestrator): extract in-loop HTTP fetch tool
httpRequestTool is the agent loop's outbound-network surface — the
only network egress the agent has by default beyond MCP calls. It's
self-contained: its own request struct (httpRequestArgs), policy
gating (HTTPRequestPolicy on the executor), max-bytes / timeout
enforcement, and a single Executor.Execute-shaped return. Splitting
it out of executor_agent_loop.go is a clean cut.

  executor_agent_loop.go:      2112 → 1999 LOC
  executor_agent_loop_http.go: new, 113 LOC

What moves to executor_agent_loop_http.go

  - httpRequestArgs: the {url, method, headers, body} JSON arg
    shape the LLM passes via the http_request tool catalog entry.
  - (e *AgentLoopExecutor).httpRequestTool: the actual fetch +
    policy gate + response cap + step/artifact builder.

What stays in executor_agent_loop.go

  Construction, Execute loop, the in-tool-dispatch siblings
  (executor_agent_loop_tools.go), the LLM-call layer
  (executor_agent_loop_chat.go), the MCP-call telemetry helper
  (next 4d-telemetry target), the assistant-event emitters, the
  agentToolDefinitions catalog, and step/artifact builders.

The receiver type stays *AgentLoopExecutor; no exported symbol
moves; tests untouched. executor_agent_loop_http.go takes its own
context + time + pkg/types.

Verification

  go build ./...                — clean
  go vet ./...                  — clean
  gofmt -s -l .                 — clean
  go test ./... -race -count=1  — 1779 / 1779 pass
